+++ ui/easydiffusion/types.py
@@ -32,8 +32,9 @@ class TaskData(BaseModel):
     vram_usage_level: str = "balanced"  # or "low" or "medium"
 
     use_face_correction: str = None  # or "GFPGANv1.3"
-    use_upscale: str = None  # or "RealESRGAN_x4plus" or "RealESRGAN_x4plus_anime_6B"
+    use_upscale: str = None  # or "RealESRGAN_x4plus" or "RealESRGAN_x4plus_anime_6B" or "latent_upscaler"
     upscale_amount: int = 4  # or 2
+    latent_upscaler_steps: int = 10
     use_stable_diffusion_model: str = "sd-v1-4"
     # use_stable_diffusion_config: str = "v1-inference"
     use_vae_model: str = None